% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_gradients.R
\name{available_gradients}
\alias{available_gradients}
\title{Display available gradient colors}
\usage{
available_gradients()
}
\value{
a character vector
}
\description{
the function displays the available gradient colors contained within the package along with their corresponding hex colors
}
\examples{
\donttest{

available_gradients()


}



}
