% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corazon_gradient.R
\docType{data}
\name{gradient}
\alias{gradient}
\title{colorffy gradient color data set}
\format{
A data frame containing all the the gradient colors:
\itemize{
\item g_name: the name of the gradients as displayed within the colorffy website.
\item hex1: the first hex color code
\item hex2: the second hex color code
\item hex3: the third hex color code
}
}
\usage{
gradient
}
\description{
A data set containing the gradient colors available at the colorffy website
}
\keyword{datasets}
